# docs/swagger.yaml
openapi: 3.0.0
info:
  title: ITBEES Global API
  version: 1.0.0
  description: Production-grade API for ITBEES Global platform.
servers:
  - url: http://localhost:5000/api
    description: Development server
paths:
  /public/courses:
    get:
      summary: Get all courses
      responses:
        '200':
          description: A list of courses
  /public/purchase/otp:
    post:
      summary: Request OTP for purchase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
  /admin/login:
    post:
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
  /admin/forgot-password:
    post:
      summary: Request admin password reset OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
  /admin/verify-reset-otp:
    post:
      summary: Verify admin password reset OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                otp:
                  type: string
  /admin/reset-password:
    post:
      summary: Reset admin password using OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                otp:
                  type: string
                newPassword:
                  type: string
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT